Move geckolib/properties.mako.rs to style/properties/gecko.mako.rs
