+++ components/style/lib.rs
@@ -83,6 +83,7 @@ pub mod error_reporting;
 pub mod font_face;
 #[cfg(feature = "gecko")] pub mod gecko_conversions;
 #[cfg(feature = "gecko")] pub mod gecko_values;
+#[cfg(feature = "gecko")] pub mod gecko_glue;
 pub mod keyframes;
 pub mod logical_geometry;
 pub mod matching;
@@ -114,6 +115,12 @@ use std::sync::Arc;
 #[allow(unsafe_code)]
 pub mod properties {
     include!(concat!(env!("OUT_DIR"), "/properties.rs"));
+}
+
+#[cfg(feature = "gecko")]
+#[allow(unsafe_code)]
+pub mod gecko_properties {
+    include!(concat!(env!("OUT_DIR"), "/gecko_properties.rs"));
 }
 
 macro_rules! reexport_computed_values {
