+++ ports/geckolib/Cargo.toml
@@ -4,8 +4,6 @@ version = "0.0.1"
 authors = ["The Servo Project Developers"]
 license = "MPL-2.0"
 
-build = "build.rs"
-
 [lib]
 name = "geckoservo"
 path = "lib.rs"
